pyproject.toml
[tool.poetry]
name = "mcp-service"
version = "0.1.0"
description = "Example mcp-service"
authors = ["Dmitriy Kulbida <dmitriy.kulbida@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.115.11"
pydantic-settings = "^2.8.1"
dependency-injector = "^4.46.0"
aioredis = "^2.0.1"
redis = "^5.2.1"
aio-pika = "^9.5.5"
uvicorn = "^0.34.0"
httpx = { version = "^0.27.0", extras = ["http2"] }
openai = "^1.68.2"
aiogram = "^3.19.0"
yandex-cloud-ml-sdk = "^0.5.0"
markdown = "^3.7"
beautifulsoup4 = "^4.13.3"
sqlalchemy = "^2.0.29"
asyncpg = "^0.30.0"
alembic = "^1.12.1"
pymongo = "^4.12.0"
weaviate-client = "^4.14.1"


[tool.poetry.group.testing.dependencies]
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
pre-commit = "^3.8.0"
pytest-asyncio = "^0.24.0"
sqlalchemy-utils = "^0.41.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 79
target-version = ["py311"]


[tool.flake8]
max-line-length = 79


[tool.isort]
profile = "black"
line_length = 79
